Info Panel: Add column/table mapping to POST request
So the backend can use the mapping to determine
the number of tables and where the content of the
tables would be

In addition, added more instructions to indicate the
mapping is assumed from left to right of the file.
Enforce the mapping in the POST request is ordered
by column letter.