Move unique row calculations into `processFile`
In an effort to reduce the number of iterations required
to generate all the data, it would make more sense if
we calculate the unique rows when we are generating
the rows themselves.

Moreover, none of the following middleware uses
data by columns to calculate any relationships besides
data type by column, which is covered in
`res.locals.dataTypes` anyway.

This commit removes `res.locals.dataByColumns`
and adds the ability to calculate unique rows by table
in `excelController.processFile`